test(di.get.spec): split out DI.get tests into their own file
